RogueLearning.jl

Layers

RogueLearning.GenDense — Type

Generalized version of Flux's Dense layer.

GenDense(in=>out, σ=identity; 
         ω = identity, ψ = *, init = glorot_uniform, 
         bias=true, γ=Flux.Zeros()

Can also be initialized with an additional set of trainable weights

GenDense(in=>out, in_asym=>out_asym, σ = identity; 
         ω = identity, ψ = *, init = glorot_uniform, 
         bias=true, bias_asym=true, γ=Flux.Zeros())

RogueLearning.GenConv — Type

Generalized version of Flux's conv layer

Similarity/correlation functions

RogueLearning.matmul — Function

Regular matrix multiplication.

RogueLearning.matmul_asym_∂x — Function

Compute matrix multiplication, but takes an additional matrix B as input. B has same dims as Wᵀ, and is used in the backwards pass.

RogueLearning.matmul_blocked_∂x — Function

Matrix multiplication with custom rrule

RogueLearning.radialSim — Function

Compute negative squared euclidean distance D between the rows of matrix W and the columns of matrix X. Denoting the rows of W by index i and the columns of X by index j the elements of the output matrix is given by: Dᵢⱼ = -||Wᵢ﹕ - X﹕ⱼ||² = 2Wᵢ﹕X﹕,j - ||Wᵢ﹕||^2 - ||X﹕ⱼ||².

RogueLearning.radialSim_asym — Function

In the forward pass this function behaves just like radialSim, but in the backwards pass weight symmetry is broken by using matrix B rather than Wᵀ. See docstring for radialSim for more details.

RogueLearning.conv_asym_∂x — Function

computes the convolution of image x with kernel w when called, but uses a different set of weights w_asym to compute the pullback wrt x. This is typically uses in feedback alignment experiments.

Loss functions

RogueLearning.direct_feedback_loss — Function

Error function which takes a vector of the hidden and output neurons states as well as a vector of feedback matrices as arguments

Activation functions