RogueLearning.jl

Function Documentation

ChainRulesCore.rrule — Method

This rrule blocks the regular backprop pathway by returning NoTangent() instead of ∂X

ChainRulesCore.rrule — Method

When conv is called with an additional weight matrix wFB, then we will use this custom rrule, which transports error backwards using wFB instead of W.

ChainRulesCore.rrule — Method

This rrule pipes output errors directly back into the hidden layers via the feedback matrices Bᵢ. To use it use Flux.activations to get hidden layers activation states and pass the array of hidden states in as the argument x along with an array of appropriately dimensioned feedback matrices.

ChainRulesCore.rrule — Method

rrule which uses feedback weights B instead of Wᵀ, which regular backpropagation would use.

NNlib.conv — Method

conv called with an additional feedback weight matrix. The forward pass gives the same results, but in the backwards pass the gradient will be different thanks to multiple dispatch.

RogueLearning.dfa_loss — Method

Error function which takes a vector of the hidden and output neurons states as well as a vector of feedback matrices as arguments

RogueLearning.matmul — Method

Compute matrix multiplication, but takes an additional matrix B as input. B has same dims as Wᵀ, and is used in the backwards pass.

RogueLearning.matmul — Method

Regular matrix multiplication.

RogueLearning.matmul_dfa — Method

Matrix multiplication with custom rrule

RogueLearning.radialSim — Method

In the forward pass this function behaves just like radialSim, but in the backwards pass weight symmetry is broken by using matrix B rather than Wᵀ. See docstring for radialSim for more details.

RogueLearning.radialSim — Method

Compute negative squared euclidean distance D between the rows of matrix W and the columns of matrix X. Denoting the rows of W by index i and the columns of X by index j the elements of the output matrix is given by: Dᵢⱼ = -||Wᵢ﹕ - X﹕ⱼ||² = 2Wᵢ﹕X﹕,j - ||Wᵢ﹕||^2 - ||X﹕ⱼ||².

RogueLearning.GenConv — Type

Generalized version of Flux's conv layer

RogueLearning.GenDense — Type

Generalized version of Flux's Dense layer.

GenDense(in=>out, σ=identity; ω = identity, ψ = *, init = glorot_uniform, bias=true, γ=Flux.Zeros()

Can also be initialized with an additional set of trainable weights asym

GenDense(in=>out, in_asym=>out_asym, σ = identity; ω = identity, ψ = *, init = glorot_uniform, bias=true, bias_asym=true, γ=Flux.Zeros())